+++ calendar-bundle/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Contao calendar bundle change log
+
+## DEV
+
+ * Show the 404 page if an external event is opened in the event reader (see contao/news-bundle#33).