+++ calendar-bundle/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## DEV
 
+ * Pass the number of events to the main template (see #32).
  * Make the calendar model available in the event reader (see contao/core#8869).
  * Show the 404 page if an external event is opened in the event reader (see contao/news-bundle#33).